Set 2 cards in A3
For printing of regular sized greeting cards, it is required to resize
and fit two cards into one A3 sheet. The program resizes and fits the
front and back side of greeting card layers that is placed on a
predefined A3 paper template. The two sides are auto saved. into Print
folder